+++ History.md
@@ -1,3 +1,5 @@
+* SeanHall: WIXFEAT:4413 - Add the WixBundleNumberOfApplyPhases built-in variable.
+
 * BobArnson: WIXBUG:4394 - Enforce a maximum include nesting depth of 1024 to avoid stack overflows when files include themselves.
 
 * johnbuuck: WIXBUG:4279 - Add support for MSBuild version 12.0.